feat: address feedback
