+++ Course/Course/Domain/Model/CourseProgressDetails.swift
@@ -52,6 +52,55 @@ public struct CourseProgressDetails: Sendable {
         self.verificationData = verificationData
         self.assignmentColors = assignmentColors
     }
+    
+    public func getAssignmentProgress(for assignmentType: String) -> AssignmentProgressData {
+        guard let policy = self.gradingPolicy.assignmentPolicies
+            .first(where: { $0.type == assignmentType }) else {
+            return AssignmentProgressData(
+                completed: 0,
+                total: 0,
+                earnedPoints: 0.0,
+                possiblePoints: 0.0,
+                percentGraded: 0.0
+            )
+        }
+        
+        let assignments = self.sectionScores.flatMap { $0.subsections }
+            .filter { $0.assignmentType == assignmentType && $0.hasGradedAssignment }
+        
+        // Calculate completed and total based on problem scores
+        var completedProblems = 0
+        var totalProblems = 0
+        
+        for assignment in assignments {
+            // Count problems in this assignment
+            totalProblems += assignment.problemScores.count
+            
+            // Count completed problems (where earned > 0)
+            completedProblems += assignment.problemScores.filter { $0.earned > 0 }.count
+        }
+        
+        // If no problem scores available, fall back to subsection-based counting
+        if totalProblems == 0 {
+            completedProblems = assignments.filter { $0.numPointsEarned > 0 }.count
+            totalProblems = policy.numTotal
+        }
+        
+        let earnedPoints = assignments.reduce(0.0) { $0 + $1.numPointsEarned }
+        let possiblePoints = assignments.reduce(0.0) { $0 + $1.numPointsPossible }
+        
+        // Calculate average percent_graded for this assignment type (from server data)
+        let totalPercentGraded = assignments.reduce(0.0) { $0 + $1.percentGraded }
+        let averagePercentGraded = assignments.isEmpty ? 0.0 : totalPercentGraded / Double(assignments.count)
+        
+        return AssignmentProgressData(
+            completed: completedProblems,
+            total: totalProblems,
+            earnedPoints: earnedPoints,
+            possiblePoints: possiblePoints,
+            percentGraded: averagePercentGraded
+        )
+    }
 }
 
 // MARK: - AssignmentProgressData
